move i18nSettings into utils and out of views
