+++ src/App.js
@@ -39,7 +39,7 @@ import UsNdFacilitiesExplore from "./views/tenants/us_nd/facilities/Explore";
 import UsNdProgrammingExplore from "./views/tenants/us_nd/programming/Explore";
 import initFontAwesome from "./utils/initFontAwesome";
 import initIntercomSettings from "./utils/initIntercomSettings";
-import { initI18n } from "./views/tenants/utils/i18nSettings";
+import { initI18n } from "./utils/i18nSettings";
 import { LANTERN_TENANTS } from "./RootStore/TenantStore/lanternTenants";
 import { CORE_TENANTS } from "./RootStore/TenantStore/coreTenants";
 import AuthWall from "./AuthWall";
