+++ src/components/charts/new_revocations/RevocationCharts/RevocationCharts.js
@@ -20,7 +20,7 @@ import cn from "classnames";
 import PropTypes from "prop-types";
 import { observer } from "mobx-react-lite";
 
-import { translate } from "../../../../views/tenants/utils/i18nSettings";
+import { translate } from "../../../../utils/i18nSettings";
 import flags from "../../../../flags";
 import RevocationsByRiskLevel from "../RevocationsByRiskLevel/RevocationsByRiskLevel";
 import RevocationsByOfficer from "../RevocationsByOfficer";
