+++ src/components/charts/new_revocations/RevocationCountChart.js
@@ -19,7 +19,7 @@ import React from "react";
 import { Bar } from "react-chartjs-2";
 import PropTypes from "prop-types";
 
-import { translate } from "../../../views/tenants/utils/i18nSettings";
+import { translate } from "../../../utils/i18nSettings";
 import { standardTooltipForCountMetric } from "../../../utils/charts/toggles";
 import { COLORS } from "../../../assets/scripts/constants/colors";
