+++ src/components/charts/new_revocations/RevocationsByDistrict/createGenerateChartData.js
@@ -25,7 +25,7 @@ import toInteger from "lodash/fp/toInteger";
 import orderBy from "lodash/fp/orderBy";
 import { calculateRate } from "../helpers/rate";
 
-import { translate } from "../../../../views/tenants/utils/i18nSettings";
+import { translate } from "../../../../utils/i18nSettings";
 import { applyStatisticallySignificantShading } from "../../../../utils/charts/significantStatistics";
 import { COLORS } from "../../../../assets/scripts/constants/colors";
 import { sumCounts } from "../utils/sumCounts";
