+++ src/components/charts/new_revocations/RevocationsByRace/__tests__/createGenerateChartData.test.js
@@ -15,7 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { setTranslateLocale } from "../../../../../views/tenants/utils/i18nSettings";
+import { setTranslateLocale } from "../../../../../utils/i18nSettings";
 import * as lanternTenant from "../../../../../RootStore/TenantStore/lanternTenants";
 import { generateDatasets } from "../createGenerateChartData";
