+++ src/components/charts/new_revocations/RevocationsByViolation/RevocationsByViolation.js
@@ -22,7 +22,7 @@ import { observer } from "mobx-react-lite";
 import RevocationsByDimension from "../RevocationsByDimension";
 import createGenerateChartData from "./createGenerateChartData";
 import BarChartWithLabels from "../BarCharts";
-import { translate } from "../../../../views/tenants/utils/i18nSettings";
+import { translate } from "../../../../utils/i18nSettings";
 import { useRootStore } from "../../../../StoreProvider";
 import { VIOLATION_TYPE } from "../../../../constants/filterTypes";
 import { COLORS } from "../../../../assets/scripts/constants/colors";
