+++ src/components/charts/new_revocations/utils/getLabelByMode.js
@@ -15,7 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { translate } from "../../../../views/tenants/utils/i18nSettings";
+import { translate } from "../../../../utils/i18nSettings";
 import { genderValueToLabel } from "../../../../utils/transforms/labels";
 
 const getLabelByMode = (mode) => {
