+++ src/setupTests.js
@@ -19,6 +19,6 @@ import "@testing-library/jest-dom/extend-expect";
 import "@testing-library/react-hooks/dont-cleanup-after-each";
 import "jest-canvas-mock";
 
-import { initI18n } from "./views/tenants/utils/i18nSettings";
+import { initI18n } from "./utils/i18nSettings";
 
 initI18n();
