+++ src/utils/__tests__/i18nSettings.test.js
